Don't run calicoctl on node with HostEndpoint
When testing with a host endpoint and default-deny policy, it doesn't
work to run the calicoctl pod on the node with the host endpoint - most
importantly when the intention is to delete the host endpoint!  If the
calicoctl pod does try to run on that node, it fails, the host endpoint
is left in place, and then lots of subsequent tests fail.

Fix that by avoiding the host endpoint node when scheduling calicoctl.